feat(ai): implement basic ai state machine
