+++ WorkingTitle/Assets/WorkingTitle.Unity/Gameplay/Ai/AiSearchStateComponent.cs
@@ -1,0 +1,9 @@
+﻿using Sirenix.OdinInspector;
+
+namespace WorkingTitle.Unity.Gameplay.Ai
+{
+    public class AiSearchStateComponent : AiStateComponent
+    {
+        
+    }
+}
