+++ WorkingTitle/Assets/WorkingTitle.Unity/Gameplay/Ai/AiStateComponent.cs.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: f834f41020e1442abfda5d773db08795
+timeCreated: 1667817765
