--- WorkingTitle/Assets/WorkingTitle.Unity/Gameplay/AiComponent.cs
@@ -1,9 +1,0 @@
-﻿using Sirenix.OdinInspector;
-
-namespace WorkingTitle.Unity.Gameplay
-{
-    public class AiComponent : SerializedMonoBehaviour
-    {
-        
-    }
-}
