+++ WorkingTitle/Assets/WorkingTitle.Unity/StateComponent.cs.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: 9b20386639fc4ef68166a3b4cffb800b
+timeCreated: 1667818515
